App atualizando e deletando, agora temos que criar a entidade user e adicionar os atributos que estão faltando a entidade eventos.

+++ src/main/java/com/goparty/services/EventosRepository.java
@@ -2,6 +2,6 @@ package com.goparty.services;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventosRepository extends JpaRepository<Eventos, Long> {
+public interface EventosRepository extends JpaRepository<Eventos, Integer> {
 
 }
